Refactor pub/sub and executor API
